Rename celery config app to avoid package name clash
