+++ alliance_auth/__init__.py
@@ -3,4 +3,4 @@ from __future__ import absolute_import
 
 # This will make sure the app is always imported when
 # Django starts so that shared_task will use this app.
-from .celery import app as celery_app  # noqa
+from .celeryapp import app as celery_app  # noqa
