+++ services/modules/openfire/tasks.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals
 
 import logging
 
-from alliance_auth.celery import app
+from alliance_auth.celeryapp import app
 from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist
 from django.contrib.auth.models import User
